+++ cmn-data-svc-client/src/main/java/org/acumos/cds/client/CommonDataServiceRestClientMockImpl.java
@@ -1471,6 +1471,13 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 		return this.portalSolutions;
 	}
 
+	@Override
+	public RestPageResponse<MLPSolution> findPortalSolutionsByKwAndTags(String[] keywords, boolean active,
+			String[] userIds, String[] accessTypeCodes, String[] modelTypeCodes, String[] allTags, String[] anyTags,
+			RestPageRequest pageRequest) {
+		return this.portalSolutions;
+	}
+
 	public void setUserSolutions(RestPageResponse<MLPSolution> solutions) {
 		this.userSolutions = solutions;
 	}