+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/MockClientTest.java
@@ -448,6 +448,10 @@ public class MockClientTest {
 		client.setPortalSolutions(portalSols);
 		Assert.assertTrue(portalSols == client.findPortalSolutions(null, null, true, null, null, null, null, null, null,
 				null, pageRequest));
+		Assert.assertTrue(
+				portalSols == client.findPortalSolutionsByKw(null, true, null, null, null, null, pageRequest));
+		Assert.assertTrue(portalSols == client.findPortalSolutionsByKwAndTags(null, true, null, null, null, null, null,
+				pageRequest));
 
 		RestPageResponse<MLPSolution> userPrivSols = new RestPageResponse<>();
 		client.setUserSolutions(userPrivSols);